+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `source-filesystem` param `data` to add default metadata (extra) for each entry.
+- Tests for source-filesystem.
+- Tests for fetcher.
+### Fixed
+- Issue with `fetcher` function `getPostsByFilter` loading all entries when filter returned no results.
 
 ## [3.1.3] - 2021-07-22
 ### Fixed